Added several algorithms.
Added:
- Insertion sort
- Bubble sort
- std::sort
- std::stable_sort
- std::sort_heap

+++ avk/internal/common.h
@@ -14,8 +14,8 @@
 using sint = int32_t;
 using uint = uint32_t;
 #else
-using sint = ptrdiff_t;
-using uint = size_t;
+using sint = int32_t;
+using uint = uint32_t;
 #endif
 
 template <typename R, typename... P>
